+++ src/utils.js
@@ -1,4 +1,33 @@
+import NativeModule from 'module';
+
 const MODULE_TYPE = 'css/mini-extract';
+
+function findModuleById(compilation, id) {
+  const { modules, chunkGraph } = compilation;
+
+  for (const module of modules) {
+    const moduleId =
+      typeof chunkGraph !== 'undefined'
+        ? chunkGraph.getModuleId(module)
+        : module.id;
+
+    if (moduleId === id) {
+      return module;
+    }
+  }
+
+  return null;
+}
+
+function evalModuleCode(loaderContext, code, filename) {
+  const module = new NativeModule(filename, loaderContext);
+
+  module.paths = NativeModule._nodeModulePaths(loaderContext.context); // eslint-disable-line no-underscore-dangle
+  module.filename = filename;
+  module._compile(code, filename); // eslint-disable-line no-underscore-dangle
+
+  return module.exports;
+}
 
 function compareIds(a, b) {
   if (typeof a !== typeof b) {
@@ -20,4 +49,9 @@ function compareModulesByIdentifier(a, b) {
   return compareIds(a.identifier(), b.identifier());
 }
 
-export { MODULE_TYPE, compareModulesByIdentifier };
+export {
+  MODULE_TYPE,
+  findModuleById,
+  evalModuleCode,
+  compareModulesByIdentifier,
+};
